backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1 